requirements.txt
python-telegram-bot~=21.0
pandas~=2.2
reportlab~=4.2
openpyxl~=3.1
python-dotenv~=1.0
ethiopian-date-converter~=0.1
pytz~=2024.1